Dalin L 2.0: 全面清仓技术债 — Clippy 零警告 (33 files)
清理的 172 条 Clippy 警告：
- 18 处 doc comment 后多余空行
- 1 处 unused import (AttrDerive)
- 4 处 unused variable (prefix _)
- 5 处 dead_code 字段/方法 (#[allow(...)])
- 2 处 from_str_radix → str::parse
- 1 处 max/min → clamp
- 1 处 strip_prefix 手动切片修复
- 1 处 cmp() 与 Ord 混淆标注
- 1 处 len 补 is_empty
- 2 处 文档列表缩进修复
- 1 处 if 重复分支合并
- 10+ 处 auto-fix (clone on copy, unused mut, match→if let, collapse if)
- control-plane k8s controller namespace dead_code 标注

测试结果: 302 passed, 0 failed ✅
Clippy: 0 warnings ✅
All binaries build cleanly ✅

+++ pyo3-bindings/src/lib.rs
@@ -36,7 +36,7 @@ fn compile(source: &str) -> PyResult<String> {
 fn capability_leq(a: &str, b: &str) -> bool {
     let cap_a = parse_capability(a);
     let cap_b = parse_capability(b);
-    (cap_a as u8) <= (cap_b as u8)
+    cap_a <= cap_b
 }
 
 /// 将 Dalin L 能力字符串转换为内部枚举值。
